extract out ActiveReleaseAlertNotification into its own file. introduce special codepath for NotifyActiveReleaseEmailAction(NotifyEmailAction)
