+++ src/sentry/integrations/slack/message_builder/issues.py
@@ -29,8 +29,12 @@ from sentry.models import (
     Team,
     User,
 )
-from sentry.notifications.notifications.base import BaseNotification, ProjectNotification
-from sentry.notifications.notifications.rules import AlertRuleNotification, CommitData
+from sentry.notifications.notifications.active_release import CommitData
+from sentry.notifications.notifications.base import (
+    BaseNotification,
+    ProjectNotification,
+)
+from sentry.notifications.notifications.rules import AlertRuleNotification
 from sentry.notifications.utils.actions import MessageAction
 from sentry.types.integrations import ExternalProviders
 from sentry.utils import json
