+++ src/sentry/mail/notifications.py
@@ -8,8 +8,8 @@ from django.utils.encoding import force_text
 
 from sentry import options
 from sentry.models import Project, ProjectOption, Team, User
+from sentry.notifications.notifications.active_release import ActiveReleaseAlertNotification
 from sentry.notifications.notifications.base import BaseNotification, ProjectNotification
-from sentry.notifications.notifications.rules import ActiveReleaseAlertNotification
 from sentry.notifications.notify import register_notification_provider
 from sentry.types.integrations import ExternalProviders
 from sentry.utils import json
