+++ src/sentry/plugins/bases/notify.py
@@ -86,8 +86,7 @@ class NotificationPlugin(Plugin):
     def rule_notify(self, event, futures):
         rules = []
         extra = {"event_id": event.event_id, "group_id": event.group_id, "plugin": self.slug}
-        log_event = "dispatched"
-        for future in futures:
+        for future in filter(lambda f: f.rule, futures):
             rules.append(future.rule)
             extra["rule_id"] = future.rule.id
             if not future.kwargs:
@@ -100,7 +99,7 @@ class NotificationPlugin(Plugin):
         extra["project_id"] = project.id
         notification = Notification(event=event, rules=rules)
         self.notify(notification)
-        self.logger.info("notification.%s" % log_event, extra=extra)
+        self.logger.info("notification.%s" % "dispatched", extra=extra)
 
     def notify_users(self, group, event, triggering_rules, fail_silently=False, **kwargs):
         raise NotImplementedError
